Disable snippets cache for basket
As you might notice, the basket is redrawn with stale data when navigating through the browser history. That's because Naja caches all snippets into the history entry and pulls them from there when the user navigates back and forth. We want to exclude the basket snippet from this cache so that it always stays up-to-date.