Hint Naja about the target URL
The plus and minus buttons work like a charm now, but when adding a new product to basket, you might notice that Naja changes the URL to an ugly one that includes the 'do' parameter. In all similar scenarios where you'd redirect to 'this', you should instead give a hint to Naja about what the target URL should be, stripping the 'do' parameter from the query string.