Submitting a form onChange... pt. 3
But this only works when the page is freshly loaded. Because the <select> is part of a snippet, when snippets are redrawn, the instance of the select that has the listener bound to is no longer present in the DOM. Let's extract the logic to a function and call it after each and every snippet update.